==> 3. DBpedia Concepts at Different Confidence Levels.ipynb <==
import os
import pickle
import pandas as pd
import spacy
import spotlight
from collections import Counter
import time

input_dir = 'textbooks/'
output_dir = 'output/'
data_dir = 'data/'
metadata_file = 'data/metadata.csv'
text_file = 'text.pkl'
spotlight_server = 'http://192.168.99.101:2222/rest/annotate'

## Load data and spacy model

isbns = os.listdir(input_dir)

with open(os.path.join(data_dir, text_file), 'rb') as fp:
    all_text = pickle.load(fp) 
    
nlp = spacy.load('en')

## Load metadata and calculate number of pages

metadata = pd.read_csv(metadata_file, dtype = {'ISBN': 'str'})
metadata['num_pages'] = [len(all_text[isbn]) for isbn in metadata['ISBN']]
metadata

## Parse the text of each textbook

all_docs = {}

for isbn in isbns:    
    print('Parsing text for ISBN: {}'.format(isbn))
    start_page = metadata.loc[metadata['ISBN'] == isbn, 'start_page'].values[0]
    end_page = metadata.loc[metadata['ISBN'] == isbn, 'end_page'].values[0]
    
    # Extract the content portion of the textbook, and combine the pages
    text = ' '.join(all_text[isbn][(start_page-1):(end_page)])

    # Parse the textbook with spacy
    all_docs[isbn] = nlp(text)

## Use the DBpedia Spotlight server to annotate the text

def extract_spotlight_concept(annotation):
    return annotation['URI'].split('/resource/')[-1]

def spotlight_annotate_text(doc, confidence=0.5, support=1):
    # Confidence = confidence score for disambiguation / linking
    # Support = number of inlinks to the Wikipedia entry

    # Low support with high confidence
    annotations = spotlight.annotate(spotlight_server,
                                     doc.string,
                                     confidence=confidence, 
                                     support=support)
    
    # Filter out some annotations
    exclude_types = ['DBpedia:TimePeriod,DBpedia:Year']
    
    filtered_annotations = [ann for ann in annotations 
                            if ann['types'] not in exclude_types]
              
    return filtered_annotations

#### Annotate the text at different confidence levels

confidence_levels = [0.4, 0.6, 0.8]
n_top = 100

annotations_df = pd.DataFrame()

for confidence in confidence_levels:
    print('Annotating for confidence: {}'.format(confidence))
    for isbn in isbns:
        print('Annotating text for ISBN: {}'.format(isbn))
        annotations = spotlight_annotate_text(all_docs[isbn],
                                              confidence)
        
        concepts = [extract_spotlight_concept(ann) for ann in annotations]
        freq = Counter(concepts)
        common_concepts, counts = zip(*freq.most_common(n_top))
        
        annotations_df = annotations_df.append(
            pd.DataFrame({
                          'concept': list(common_concepts),
                          'count': list(counts),
                          'confidence': confidence,
                          'ISBN': isbn
                        }))         
        time.sleep(10)

#### Output common concepts to file

concepts_df = pd.merge(annotations_df, metadata, on='ISBN')

concepts_df.to_csv(os.path.join(output_dir, 'common_concepts.csv'), 
                      index=False)
